requirements.txt
streamlit
google-auth
google-auth-oauthlib
google-api-python-client
python-dateutil
pytz